\encoding{UTF-8}

\name{aicllsbc}

\alias{aicllsbc}

\title{Evaluation of model fit}

\description{
  Three methods for evaluating model fitting: 
    1. AIC (Akaike's information criterion),
    2. Log likelihood, and
    3. BIC (Bayesian information criterion, also known as Schwarz's Bayesian criterion).
}

\keyword{misc}
