\encoding{UTF-8}

\name{err_types}

\alias{err_types}

\title{To Generate Simulated Values for each PK Parameter Based on Error Types}

\description{
  Four Error Types are:
     1. Normal Error,           
     2. Uniform Error,          
     3. Normal Error*True Value, and
     4. Uniform Error*True Value
}

\keyword{misc}
