\name{fbolus.mm}

\alias{fbolus.mm}

\title{One Compartment Model IV Bolus Single Dose Michaelis-Menten Elimination}

\description{
  Consisting of user-supplied function for model definition, integration method for solving the ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithm for one compartment model IV bolus single dose 
  Michaelis-Menten elimination.
}

\keyword{misc}
