\encoding{UTF-8}

\name{fbolus.mm.SD}

\alias{fbolus.mm.SD}

\title{Fitting Functions for a One-Compartment, IV-Bolus, Single-Dose, and Michaelis-Menten Elimination Model
       for single-dosed}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a one-compartment, IV-bolus, single-dose, 
  and Michaelis-Menten elimination model.
}

\keyword{misc}
