\name{fbolus1}

\alias{fbolus1}

\title{One Compartment Model IV Bolus Single Dose}

\description{
  Consisting of user-supplied function for model definition, integration method for solving the ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithm for one compartment model IV bolus single dose.
}

\keyword{misc}
