\encoding{UTF-8}

\name{fbolus2}

\alias{fbolus2}

\title{Fitting Functions for a Two-Compartment, IV-Bolus, and Single-Dose Model}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a two-compartment, IV-bolus, and single-dose model.
}

\keyword{misc}
