\encoding{UTF-8}

\name{fbolus3.mm}

\alias{fbolus3.mm}

\title{Fitting Functions for a three-Compartment, IV-Bolus and Single-Dose Model
       with Michaelis-Menten Elimination}

\description{
  Includes user-supplied functions for model definition, integration method 
  for solving ordinary differential equations, three kinds of weighting schemes 
  for selection, and fitting algorithms for a three-compartment, IV-bolus, and
  single-dose model with Michaelis-Menten Elimination.
}

\keyword{misc}
