\name{ffirst.lagm}

\alias{ffirst.lagm}

\title{One Compartment Model Extravascular First Order Absorption with Lag Time Michaelis-Menten Elimination }

\description{
  Consisting of user-supplied function for model definition, integration method for solving the ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithm for one compartment model extravascular first order 
  absorption with lag time Michaelis-Menten elimination.
}

\keyword{misc}

