\encoding{UTF-8}

\name{ffirst.nolag}

\alias{ffirst.nolag}

\title{Fitting Functions for a One-Compartment, Extravascular, Single-Dose, and First-Ordered Absorption without Lag Time Model}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a one-compartment, extravascular, single-dose,
  and first-ordered absorption without lag time model.
}

\keyword{misc}
