\name{ffirst.nolagm}

\alias{ffirst.nolagm}

\title{Fitting Functions for a One-Compartment, Extravascular, Single-Dose, First-Ordered Absorption, 
       and Michaelis-Menten Elimination without Lag Time Model}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a one-compartment, extravascular,
  single-dose, first-ordered absorption, and Michaelis-Menten elimination without lag time model. 
}

\keyword{misc}