\name{ffirst2}

\alias{ffirst2}

\title{Two Compartment Model Extravascular First Order Absorption without Lag Time}

\description{
  Consisting of user-supplied function for model definition, integration method for solving the ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithm for two compartment model extravascular first order
  absorption without lag time.
}

\keyword{misc}