\name{finfu.mm}

\alias{finfu.mm}

\title{One Compartment Model IV Infusion with Michaelis-Menten Elimination}

\description{
  Consisting of user-supplied function for model definition, integration method for solving the ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithm for one compartment model IV infusion single dose 
  with Michaelis-Menten elimination 
}

\keyword{misc}
