\name{finfu2}

\alias{finfu2}

\title{Two Compartment Model IV Infusion Single Dose }

\description{
  Consisting of user-supplied function for model definition, integration method for solving the ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithm for two compartment model IV infusion single dose. 
}

\keyword{misc}
