\name{fmacro.one}

\alias{fmacro.one}

\title{Fitting Functions for a One-Exponential Term Model}

\description{
  Includes user-supplied functions for model definitions, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a one-exponential term model.
}

\keyword{misc}
