\name{fmacro.three}

\alias{fmacro.three}

\title{Fitting Functions for a Three-Exponential Term Model}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a three-exponential term model. 
  terms.
}

\keyword{misc}
