\name{fzero.nolagm}

\alias{fzero.nolagm}

\title{One Compartment Model Extravascular Zero Order Absorption without Lag Time Michaelis-Menten Elimination}

\description{
  Consisting of user-supplied function for model definition, integration method for solving the ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithm for one compartmrnt model extravascular zero order 
  absorption without lag time Michaelis-Menten elimination.
}

\keyword{misc}