\name{iv.route}

\alias{iv.route}

\title{Options for Parenteral Models}

\description{
  Provide four parenteral models for selection:
  one compartment model IV bolus single dose,
  one compartment model IV bolus single dose Michaelis-Menten elimination,
  one compartment model IV infusion single dose,
  one compartment model IV infusion single dose Michaelis-Menten elimination.
}

\keyword{misc}
