\encoding{UTF-8}

\name{macro}

\alias{macro}

\title{Options for Macroconstant Exponential Models}

\description{
  Provide the following three PK models for selection:
   1. one-exponential term model;
   2. two-exponential term model; and
   3. three-exponential term model.
}

\keyword{misc}