\encoding{UTF-8}

\name{noniv.route}

\alias{noniv.route}

\title{Options for Extravascular Models}

\description{
  Provide the following six PK models for selection:
   1. one-compartment, single-dose, and first-ordered absorption with lag time model;
   2. one-compartment, single-dose, and first-ordered absorption without lag time model;
   3. one-compartment, single-dose, and zero-ordered absorption without lag time model;
   4. one-compartment, single-dose, first-ordered absorption, and Michaelis-Menten elimination with lag time model;
   5. one-compartment, single-dose, first-ordered absorption, and Michaelis-Menten elimination without lag time model; and
   6. one-compartment, single-dose, zero-ordered absorption,and Michaelis-Menten elimination without lag time model.
}

\keyword{misc}

