\encoding{UTF-8}

\name{PK.fit.SD}

\alias{PK.fit.SD}

\title{Options for PK model selection for single-dosed}

\description{
  Provide the following PK models for selection:
   1. One-Compartment PK Model, 
   2. Two-Compartment PK Model,
   3. Three-Compartment PK Model (iv bolus with 1st-ordered elim.),
   4. Three-Compartment PK Model (iv bolus with MM elim.),
   5. Macroconstant Exponential Functions.
}

\keyword{misc}
