\encoding{UTF-8}

\name{PK.sim}

\alias{PK.sim}

\title{Options for Simulation Functions}

\description{
  Provide the following four PK models for selection:
   1. one-compartment, and IV (bolus or infusion) model;
   2. one-compartment, and non IV route model; 
   3. two-compartment model; and 
   4. macroconstant exponential functions.
}

\keyword{misc}