\name{sbolus.mm}

\alias{sbolus.mm}

\title{Simulation Functions for a One-Compartment, IV-Bolus, Single-Dose, and Michaelis-Menten Elimination Model}

\description{
  Includes entering initial values of PK parameters and five error types for selection: no error,
  normal error, uniform error, normal error*true value, and uniform error*true value.
}

\keyword{misc}
