\name{sbolus1}

\alias{sbolus1}

\title{Simulation function for One Compartment Model IV Bolus Single Dose}

\description{
  Consisting of entering the parameter initial value and five error types for selection: no error,
  normal error, uniform error, normal error*true value, uniform error*true value.
}

\keyword{misc}