\encoding{UTF-8}

\name{sbolus2}

\alias{sbolus2}

\title{Simulation Functions for a Two-Compartment, IV-Bolus, and Single-Dose Model}

\description{
  Includes entering initial values of PK parameters and five error types for selection: no error,
  normal error, uniform error, normal error*true value, and uniform error*true value.
}

\keyword{misc}