\name{sfirst.lag}

\alias{sfirst.lag}

\title{Simulation Function for One Compartment Model Extravascular First Order Absorption with Lag Time}

\description{
  Consisting of entering the parameter initial value and five error types for selection: no error,
  normal error, uniform error, normal error*true value, uniform error*true value.
}

\keyword{misc}