\name{sfirst.mm.out}

\alias{sfirst.mm.out}

\title{The Output of Simulation Function for One Compartment Model Extravascular First Order Absorption Michaelis-Menten Elimination}

\description{
  Display the output of simulation for one compartment model extravascular first order absorption 
  with or without lag time michaelis-menten elimination and plot.
}

\keyword{misc}