\encoding{UTF-8}

\name{sfirst.nolag}

\alias{sfirst.nolag}

\title{Simulation Functions for a One-Compartment, Extravascular, Single-Dose, and First-Ordered Absorption without Lag Time Model}

\description{
  Includes entering initial values of PK parameters and five error types for selection: no error,
  normal error, uniform error, normal error*true value, uniform error*true value.
}

\keyword{misc}