\encoding{UTF-8}

\name{sfirst.nolagm}

\alias{sfirst.nolagm}

\title{Simulation Functions for a One-Compartment, Extravascular, Single-Dose, First-Ordered Absorption,
       and Michaelis-Menten Elimination without Lag Time Model}

\description{
  Consisting of entering the parameter initial value and five error types for selection: no error,
  normal error, uniform error, normal error*true value, uniform error*true value.
}

\keyword{misc}