\name{sfirst1.out}

\alias{sfirst1.out}

\title{The Output of Simulation for One Compartment Model Extravascular First Order Absorption}

\description{
  Display the output of simulation for one compartment model extravascular single dose first order absorption 
  with or without lag time and plot.
}

\keyword{misc}