\name{sfirst2.out}

\alias{sfirst2.out}

\title{The Output of Simulation for Two Compartment Model Extravascular First Order Absorption without Lag Time}

\description{
  Display the output of simulation for two compartment model extravascular single dose first order absorption 
  without lag time and plot.
}

\keyword{misc}