\name{sinfu.mm.out}

\alias{sinfu.mm.out}

\title{The Output of Simulation for One Compartment Model IV Infusion Single Dose Michaelis-Menten Elimination}

\description{
  Display the output of simulation for one compartment model IV infusion single dose Michaelis-Menten elimination and plot.
}

\keyword{misc}
