\name{sinfu.mm}

\alias{sinfu.mm}

\title{Simulation function for One Compartment Model IV Infusion with Michaelis-Menten Elimination }

\description{
  Consisting of entering the parameter initial value and five error types for selection: no error,
  normal error, uniform error, normal error*true value, uniform error*true value.
}

\keyword{misc}