\name{sinfu1}

\alias{sinfu1}

\title{Simulation Functions for a One-Compartment, IV-Infusion, and Single-Dose Model}

\description{
  Includes entering initial values of PK parameters and five error types for selection: no error,
  normal error, uniform error, normal error*true value, and uniform error*true value.
}

\keyword{misc}
