\name{smacro.one}

\alias{smacro.one}

\title{Simulation Function for One Exponential Term}

\description{
  Consisting of entering the parameter initial value and five error types for selection: no error,
  normal error, uniform error, normal error*true value, uniform error*true value.
}

\keyword{misc}
