\name{smacro}

\alias{smacro}

\title{Options for Macroconstant Exponential Models Simulation}

\description{
  Provides the following three models for selection:
    1. one-exponential term model;
    2. two-exponential term model; and
    3. three-exponential term model.
}

\keyword{misc}