\name{smacro.three}

\alias{smacro.three}

\title{Simulation Function for Three Exponential Terms}

\description{
  Consisting of entering the parameter initial value and five error types for selection: no error,
  normal error, uniform error, normal error*true value, uniform error*true value.
}

\keyword{misc}