\name{sone.noniv.route}

\alias{sone.noniv.route}

\title{Option for One-Compartment, and Extravascular PK Models Simulation}

\description{
  Provide the following six PK models for selection:
    1. single-dose, and first-ordered absorption with lag time model;
    2. single-dose, and first-ordered absorption without lag time model;
    3. single-dose, and zero-ordered absorption without lag time model;
    4. single-dose, first-ordered absorption, and Michaelis-Menten elimination with lag time model;
    5. single-dose, first-ordered absorption, and Michaelis-Menten elimination without lag time model; and
    6. single-dose, zero-ordered absorption, and Michaelis-Menten elimination without lag time model.
}

\keyword{misc}
