\encoding{UTF-8}

\name{stwo.all}

\alias{stwo.all}

\title{Options for Two-Compartment PK Models Simulation}

\description{
  Provides the following three models for selection:
    1. two-compartment, IV-bolus, and single-dose model;
    2. two-compartment, IV-infusion, and single-dose model; and
    3. two-compartment, extravascular, single-dose, and first-ordered absorption without lag time model.
}

\keyword{misc}
