\name{szero.mm.out}

\alias{szero.mm.out}

\title{The Output of Simulation for Extravascular Single Dose Zero Order Absorption without Lag Time Michaelis-Menten Elimination}

\description{
  Display the output of simulation for one compartment model extravascular zero order absorption without lag time Michaelis-Menten elimination
}

\keyword{misc}