\encoding{UTF-8}

\name{two.list.SD}

\alias{two.list.SD}

\title{Options for Two-Compartment Models for single-dosed}

\description{
  Provide the following three PK models for selection:
   1. two-compartment, IV-bolus, and single-dose model; 
   2. two-compartment, IV-infusion, and single-dose model; and
   3. two-compartment, extravascular, single-dose, and first-ordered absorption without lag time model.
}

\keyword{misc}