\name{fbolus1}

\alias{fbolus1}

\title{Fitting Functions for a One-Compartment, IV-Bolus, and Single-Dose Model}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a one-compartment, IV-bolus, and single-dose model.
}

\keyword{misc}
