\name{ffirst.lagm}

\alias{ffirst.lagm}

\title{Fitting Functions for a One-Compartment, Extravascular, Single-Dose, First-Ordered Absorption,
       and Michaelis-Menten Elimination with Lag Time Model}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a one-compartment, extravascular, single-dose,
  first-ordered absorption, and Michaelis-Menten elimination with lag time model.
}

\keyword{misc}

