\name{ffirst.nolag}

\alias{ffirst.nolag}

\title{One Compartment Model Extravascular First Order Absoption without Lag Time}

\description{
  Consisting of user-supplied function for model definition, integration method for solving the ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithm for one compartment model extravascular first order 
  absorption without lag time
}

\keyword{misc}
