\encoding{UTF-8}

\name{ffirst2}

\alias{ffirst2}

\title{Fitting Functions for a Two-Compartment, Extravascular, and First-Ordered Absorption without Lag Time Model}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a two-compartment, extravascular, and first-ordered
  absorption without lag time model.
}

\keyword{misc}