\name{finfu.mm}

\alias{finfu.mm}

\title{Fitting Functions for a One-Compartment, IV-Infusion, Single-Dose, and Michaelis-Menten Elimination Model}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a one-compartment, IV-infusion, single-dose, 
  and Michaelis-Menten elimination model. 
}

\keyword{misc}
