\encoding{UTF-8}

\name{finfu1}

\alias{finfu1}

\title{Fitting Functions for a One-Compartment, IV-Infusion, and Single-Dose Model}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a one-compartment, IV-infusion, and single-dose model. 
}

\keyword{misc}