\name{finfu2}

\alias{finfu2}

\title{Fitting Functions for a Two-Compartment, IV-Infusion, and Single-Dose Model}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a two-compartment, IV-infusion, and single-dose model. 
}

\keyword{misc}
