\name{fmacro.one}

\alias{fmacro.one}

\title{One Exponential Term}

\description{
  Consisting of user-supplied function for model definition, integration method for solving the ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithm for macroconstant exponential function: one exponential 
  term.
}

\keyword{misc}
