\name{fmacro.three}

\alias{fmacro.three}

\title{Three Exponential Terms}

\description{
  Consisting of user-supplied function for model definition, integration method for solving the ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithm for macroconstant exponential function: three exponential 
  terms.
}

\keyword{misc}
