\name{fmacro.two}

\alias{fmacro.two}

\title{Fitting Functions for a Two-Exponential Term Model}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a two-exponential term model.
}

\keyword{misc}