\encoding{UTF-8}

\name{fzero.nolagm}

\alias{fzero.nolagm}

\title{Fitting Functions for a One-Compartment, Extravascular, Single-Dose, Zero-Ordered Absorption,
       and Michaelis-Menten Elimination without Lag Time Model}

\description{
  Includes user-supplied functions for model definition, integration method for solving ordinary differential equations,
  three kinds of weighting schemes for selection, and fitting algorithms for a one-compartment, extravascular,
  single-dose, zero-ordered absorption, and Michaelis-Menten elimination without lag time model.
}

\keyword{misc}