\name{iv.route}

\alias{iv.route}

\title{Options for Parenteral Models}

\description{
  Provide the following four PK models for selection:
   1. one-compartment, IV-bolus, and single-dose model;
   2. one-compartment, IV-bolus, single-dose, and Michaelis-Menten elimination model;
   3. one-compartment, IV-infusion, and single dose model; and
   4. one-compartment, IV-infusion, single dose, and Michaelis-Menten elimination model.
}

\keyword{misc}
