\name{noniv.route}

\alias{noniv.route}

\title{Options for Extravascular Models}

\description{
  Provide six extravascular models for selection:
  one compartment model single dose first-order absorption with lag time,
  one compartment model single dose first-order absorption without lag time,
  one compartment mldel single dose zero-order absorption without lag time,
  one compartment model single dose first-order absorption with lag time michaelis-menten elimination,
  one compartment model single dose first-order absorption without lag time michaelis-menten elimination,
  one compartment modle single dose zero-order absorption without lag time michaelis-menten elimination.
}

\keyword{misc}

