\name{sbolus.mm.out}

\alias{sbolus.mm.out}

\title{The Output of Simulation for One Compartment Model IV Bolus Single Dose Michaelis-Menten Elimination}

\description{
  Display the output of simulation for one compartment model IV bolus single dose michaelis-menten elimination and plot.
}

\keyword{misc}