\name{sbolus.mm}

\alias{sbolus.mm}

\title{Simulation Function for One Compartment Model IV Bolus Michaelis-Menten Elimination}

\description{
  Consisting of entering the parameter initial value and five error types for selection: no error,
  normal error, uniform error, normal error*true value, uniform error*true value.
}

\keyword{misc}
