\encoding{UTF-8}

\name{sbolus1}

\alias{sbolus1}

\title{Simulation functions for a One-Compartment, IV-Bolus, and Single-Dose Model}

\description{
  Includes entering initial values of PK parameters and five error types for selection: no error,
  normal error, uniform error, normal error*true value, and uniform error*true value.
}

\keyword{misc}