\encoding{UTF-8}

\name{sfirst.lag}

\alias{sfirst.lag}

\title{Simulation Functions for a One-Compartment, Extravascular, Single-Dose, and First-Ordered Absorption with Lag Time Model}

\description{
  Includes entering initial values Of PK parameters and five error types for selection: no error,
  normal error, uniform error, normal error*true value, and uniform error*true value.
}

\keyword{misc}