\encoding{UTF-8}

\name{sfirst.lagm}

\alias{sfirst.lagm}

\title{Simulation Functions for a One-Compartment, Extravascular, Single-Dose, First-Ordered Absorption,
      and Michaelis-Menten Elimination with Lag Time Model}

\description{
  Includes entering initial values of PK parameters and five error types for selection: no error,
  normal error, uniform error, normal error*true value, and uniform error*true value.
}

\keyword{misc}