\name{sfirst.mm.out}

\alias{sfirst.mm.out}

\title{Simulation Output for a One-Compartment, Extravascular, First-Ordered Absorption, and Michaelis-Menten Elimination Model}

\description{
  Display simulation output for a one-compartment, extravascular, first-ordered absorption,
  and Michaelis-Menten elimination model and plots.
}

\keyword{misc}