\name{sfirst1.out}

\alias{sfirst1.out}

\title{Simulation Output for a One-Compartment, Extravascular, Single-Dose, and First-Ordered Absorption Model}

\description{
  Display simulation output for a one-compartment, extravascular, single-dose, and first-ordered absorption model and plots. 

}

\keyword{misc}