\encoding{UTF-8}

\name{sfirst2.out}

\alias{sfirst2.out}

\title{Simulation Output for a Two-Compartment, Extravascular, Single-Dose, and First-Ordered Absorption without Lag Time Model}

\description{
  Display simulation output for a two-compartment, extravascular, single-dose, and first-ordered absorption 
  without lag time model and plots.
}

\keyword{misc}