\name{sfirst2}

\alias{sfirst2}

\title{Simulation Functions for a Two-Compartment, Extravascular, Single-Dose, and First-Ordered Absorption without Lag Time Model}

\description{
  Includes entering initial values of PK parameters and five error types for selection: no error,
  normal error, uniform error, normal error*true value, and uniform error*true value.
}

\keyword{misc}