\name{sinfu.mm.out}

\alias{sinfu.mm.out}

\title{Simulation Output for a One-Compartment, IV-Infusion, Single-Dose, and Michaelis-Menten Elimination Model}

\description{
  Display simulation output for a one-compartment, IV-infusion, single-dose and Michaelis-Menten elimination model and plots.
}

\keyword{misc}
