\name{sinfu1}

\alias{sinfu1}

\title{Simulation Function for One Compartment Model IV infusion Single Dose}

\description{
  Consisting of entering the parameter initial value and five error types for selection: no error,
  normal error, uniform error, normal error*true value, uniform error*true value.
}

\keyword{misc}
