\name{sinfu2}

\alias{sinfu2}

\title{Simulation Function for Two Compartment Model IV infusion Single Dose}

\description{
  Consisting of entering the parameter initial value and five error types for selection: no error,
  normal error, uniform error, normal error*true value, uniform error*true value.
}

\keyword{misc}