\name{smacro.one}

\alias{smacro.one}

\title{Simulation Functions for a One-Exponential Term Model}

\description{
  Includes entering initial values of PK parameters and five error types for selection: no error,
  normal error, uniform error, normal error*true value, and uniform error*true value.
}

\keyword{misc}
