\encoding{UTF-8}

\name{smacro.three}

\alias{smacro.three}

\title{Simulation Functions for a Three-Exponential Term Model}

\description{
  Includes entering initial values of PK parameters and five error types for selection: no error,
  normal error, uniform error, normal error*true value, and uniform error*true value.
}

\keyword{misc}