\name{sone.iv.route}

\alias{sone.iv.route}

\title{Options for Parenteral Simulation}

\description{
  Provide four models for selection: 
  IV bolus single dose, 
  IV bolus single dose nonlinear elimination, 
  IV infusion single dose,
  IV infusion single dose nonlinear elimination.
}

\keyword{misc}