\name{sone.noniv.route}

\alias{sone.noniv.route}

\title{Option for Extravascular Models Simulation}

\description{
  Provide six models for selection:
  single dose first-order absorption with lag time,
  single dose first-order absorption without lag time,
  single dose zero-order absorption without lag time,
  single dose first-order absorption with lag time nonlinear elimination,
  single dose first-order absorption without lag time nonlinear elimination,
  single dose zero-order absorption without lag time nonlinear elimination
}

\keyword{misc}
