\name{stwo.all}

\alias{stwo.all}

\title{Options for Two Compartment Model Simulation}

\description{
  Provide three models for selection:
  two compartment model IV bolus single dose, 
  two compartment model IV infusion single dose,
  two compartment model extravascular single dose first-order absorption without lag time.
}

\keyword{misc}
