\name{szero.mm.out}

\alias{szero.mm.out}

\title{Simulation Output for a One-Compartment, Extravascular, Single-Dose, Zero-Ordered Absorption, 
       and Michaelis-Menten Elimination without Lag Time Model}

\description{
  Display simulation output for a one-compartment, extravascular, single-dose, zero-ordered absorption, 
  and Michaelis-Menten elimination without lag time model and plots.
}

\keyword{misc}