\name{szero.out}

\alias{szero.out}

\title{The Output of Simulation for One Compartment Model Extravascular Single Dose Zero Order Absorption without Lag Time}

\description{
  Display the output of simulation for one compartment model extravascular single dose zero order absorption without lag time
}

\keyword{misc}